Apply (most) IWYU recommendataions to include and src
Reduces aggregate compilation unit header usage, and is explicit about
dependencies.

Other changes included:
- move everything in `types.hpp` to other headers and dropped
  `types.hpp`.
- removed the `concepts` namespace
- replace trivial concepts (stream_derived_type) with their inline
  versions (std::derived_from<Stream>).
