+++ include/oxen/quic.hpp
@@ -8,19 +8,18 @@
 #include "quic/crypto.hpp"
 #include "quic/datagram.hpp"
 #include "quic/endpoint.hpp"
-// #include "quic/format.hpp"  // Not included by default as it depends on fmt
-#include "quic/error.hpp"
-#include "quic/format.hpp"
 #include "quic/formattable.hpp"
-#include "quic/gnutls_crypto.hpp"
 #include "quic/iochannel.hpp"
 #include "quic/ip.hpp"
 #include "quic/loop.hpp"
 #include "quic/messages.hpp"
 #include "quic/network.hpp"
 #include "quic/opt.hpp"
+#include "quic/result.hpp"
 #include "quic/stream.hpp"
-#include "quic/types.hpp"
 #include "quic/udp.hpp"
 #include "quic/utils.hpp"
 #include "quic/version.hpp"
+
+// #include "quic/format.hpp"  // Not included by default as it depends on fmt
+// #include "quic/gnutls_crypto.hpp"  // Not included by default as it depends on gnutls
