+++ include/oxen/quic/connection_ids.hpp
@@ -1,7 +1,11 @@
 #pragma once
 
-#include "formattable.hpp"
-#include "types.hpp"
+#include <ngtcp2/ngtcp2.h>
+
+#include <cassert>
+#include <cstdint>
+#include <cstring>
+#include <string>
 
 namespace oxen::quic
 {
