+++ include/oxen/quic/network.hpp
@@ -3,14 +3,16 @@
 #include "loop.hpp"
 
 #include <atomic>
-#include <cstdint>
-#include <future>
+#include <chrono>
+#include <functional>
 #include <memory>
-#include <thread>
+#include <unordered_set>
+#include <utility>
 
 namespace oxen::quic
 {
     class Endpoint;
+    struct Address;
 
     /** Network:
             This object is the entry point to libquic, providing functionalities like job scheduling and endpoint creation
