+++ include/oxen/quic/result.hpp
@@ -2,6 +2,16 @@
 
 #include "utils.hpp"
 
+#include <ngtcp2/ngtcp2.h>
+
+#include <cstdint>
+#include <exception>
+#include <string>
+
+#ifdef _WIN32
+#include <array>
+#endif
+
 namespace oxen::quic
 {
     // Maximum allowed application error code; this is defined by the QUIC protocol itself, that
@@ -55,32 +65,66 @@ namespace oxen::quic
     // Stateless reset received
     inline constexpr uint64_t CONN_STATELESS_RESET = ERROR_BASE + 1005;
 
-    inline std::string quic_strerror(uint64_t e)
+    std::string quic_strerror(uint64_t e);
+
+    struct ngtcp2_error_code_t final
+    {};
+
+    // Tag value to pass into the io_result constructor to indicate an ngtcp2 error code.  (For
+    // ngtcp2, error codes are arbitrary negative values without any connection to errno).
+    inline constexpr ngtcp2_error_code_t ngtcp2_error_code{};
+
+    // Struct returned as a result of send_packet that either is implicitly
+    // convertible to bool, but also is able to carry an error code
+    struct io_result
     {
-        switch (e)
+        // Default construction makes a "good" io_result, i.e. with error code 0
+        io_result() : io_result{0} {}
+
+        // Constructs an io_result with an `errno` value.
+        explicit io_result(int errno_val) : error_code{errno_val} {}
+
+        // Constructs an io_result with an ngtcp2 error value.
+        io_result(int err, ngtcp2_error_code_t) : error_code{err}, is_ngtcp2{true} {}
+
+#ifdef _WIN32
+        static io_result wsa(int err)
         {
-            case 0:
-                return "No error"s;
-            case DATAGRAM_ERROR_EXCEPTION:
-                return "Error - datagram exception"s;
-            case STREAM_ERROR_EXCEPTION:
-                return "Error - stream exception"s;
-            case BPARSER_ERROR_EXCEPTION:
-                return "Error - bt request stream exception"s;
-            case STREAM_ERROR_CONNECTION_CLOSED:
-                return "Error - stream connection closed"s;
-            case CONN_WRITE_CLOSE_FAIL:
-                return "Error - Failed to write connection close"s;
-            case CONN_SEND_CLOSE_FAIL:
-                return "Error - Failed to send connection close"s;
-            case CONN_SEND_FAIL:
-                return "Error - Failed to send packet"s;
-            case CONN_IDLE_CLOSED:
-                return "Connection closed by idle timeout"s;
-            default:
-                return "Application error code " + std::to_string(e);
+            io_result e{err};
+            e.is_wsa = true;
+            return e;
         }
-    }
+#endif
+
+        // Same as the ngtcp2 error code constructor
+        static io_result ngtcp2(int err) { return io_result{err, ngtcp2_error_code}; }
+
+        // The numeric error code
+        int error_code{0};
+        // If true then `error_code` is an ngtcp2 error code, rather than an errno value.
+        bool is_ngtcp2 = false;
+#ifdef _WIN32
+        // If true then this is a WSALastErrorCode error code value.
+        bool is_wsa = false;
+#endif
+        // Returns true if this indicates success, i.e. error code of 0
+        bool success() const { return error_code == 0; }
+        // Returns true if this indicates failure, i.e. error code not 0
+        bool failure() const { return !success(); }
+        // returns true if error value indicates a failure to write without blocking
+        bool blocked() const
+        {
+            return is_ngtcp2 ? error_code == NGTCP2_ERR_STREAM_DATA_BLOCKED
+#ifdef _WIN32
+                 : is_wsa ? error_code == WSAEWOULDBLOCK
+#endif
+                          : (error_code == EAGAIN || error_code == EWOULDBLOCK);
+        }
+
+        // returns the error message string describing error_code
+        std::conditional_t<IN_HELL, std::string, std::string_view> str_error() const;
+        
+    };
 
     struct io_error
     {
@@ -103,7 +147,7 @@ namespace oxen::quic
 
         uint64_t ngtcp2() const;
 
-        std::string strerror() const { return is_ngtcp2 ? ngtcp2_strerror(static_cast<int>(_code)) : quic_strerror(_code); }
+        std::string strerror() const;
     };
 
 }  // namespace oxen::quic
