+++ include/oxen/quic/udp.hpp
@@ -13,11 +13,13 @@ extern "C"
 }
 
 #include "address.hpp"
-#include "types.hpp"
+#include "result.hpp"
 #include "utils.hpp"
 
 #include <cstdint>
 #include <variant>
+
+struct event_base;
 
 namespace oxen::quic
 {
@@ -47,7 +49,9 @@ namespace oxen::quic
         std::span<const Char> data() const
         {
             return std::visit(
-                    [](const auto& d) { return std::span<const Char>{reinterpret_cast<const Char*>(d.data()), d.size()}; },
+                    [](const auto& d) {
+                        return std::span<const Char>{reinterpret_cast<const Char*>(d.data()), d.size()};
+                    },
                     pkt_data);
         }
