+++ src/endpoint.cpp
@@ -1,26 +1,37 @@
 #include "endpoint.hpp"
 
-extern "C"
-{
-#include <ngtcp2/ngtcp2.h>
-#include <ngtcp2/version.h>
-#ifdef __linux__
-#include <netinet/udp.h>
-#endif
-}
-
 #include "connection.hpp"
-#include "gnutls_crypto.hpp"
+#include "context.hpp"
 #include "internal.hpp"
 #include "opt.hpp"
-#include "types.hpp"
 #include "utils.hpp"
 
-#include <event2/event.h>
+#include <oxenc/hex.h>
 
+#include <event2/event.h>
+#include <ngtcp2/ngtcp2.h>
+#include <ngtcp2/ngtcp2_crypto.h>
+
+#include <gnutls/crypto.h>
+
+#include <algorithm>
+#include <array>
+#include <cassert>
+#include <cerrno>
 #include <cstddef>
+#include <cstring>
 #include <list>
+#include <numeric>
 #include <optional>
+#include <string_view>
+#include <tuple>
+
+#ifndef _WIN32
+extern "C"
+{
+#include <sys/time.h>
+}
+#endif
 
 namespace oxen::quic
 {
