+++ src/stream.cpp
@@ -1,20 +1,22 @@
 #include "stream.hpp"
 
-extern "C"
-{
-#include <ngtcp2/ngtcp2.h>
-}
-
 #include "connection.hpp"
-#include "context.hpp"
 #include "endpoint.hpp"
 #include "internal.hpp"
-#include "network.hpp"
-#include "types.hpp"
+#include "messages.hpp"
+#include "result.hpp"
+
+#include <ngtcp2/ngtcp2.h>
 
 #include <cstddef>
-#include <cstdio>
+#include <exception>
+#include <iterator>
+#include <optional>
+#include <span>
 #include <stdexcept>
+#include <string_view>
+#include <utility>
+#include <vector>
 
 namespace oxen::quic
 {
