+++ tests/speedtest-server.cpp
@@ -2,8 +2,11 @@
     Test server binary
 */
 
-#include "oxen/quic/opt.hpp"
 #include "utils.hpp"
+
+#include <oxen/quic/opt.hpp>
+
+#include <gnutls/crypto.h>
 
 using namespace oxen::quic;
 
@@ -50,8 +53,7 @@ int main(int argc, char* argv[])
 
     Network server_net{};
 
-    auto [listen_addr, listen_port] = parse_addr(server_addr, DEFAULT_SPEEDTEST_ADDR.port());
-    Address server_local{listen_addr, listen_port};
+    auto server_local = Address::parse(server_addr, DEFAULT_SPEEDTEST_ADDR.port());
 
     stream_open_callback stream_opened = [&](Stream& s) {
         log::warning(test_cat, "Stream {} opened!", s.stream_id());
